st-hal: fix ec ref mixer not reset
For barge in usecaes, when session is switched to non-LPI mode,
ec ref can be enabled if valid profile is set.
When the background music is stopped, lpi flag can be reset, this
makes ec profile returned for this session to be none, thus reset
of ec ref mixer is skipped.
Then next time when the session is stopped and started again on
non-LPI mode, ec won't be applied.
When getting the ec profile, check the current session lpi mode.

Change-Id: I394552059b3532877f313fb981ba57c3e7b002c1