st-hal: fix session back to active due to playback event
When capture concurrency is disabled and there's an active
capture usecase, va sessions could be put into paused, while
by starting a playback stream can bring back the sessions to
be active again, which is not expected.
Check both playback concurrency and capture concurrency to
to get the final decision for concurrency allowance.

Change-Id: Iac9e2b94121d45cd6d93c7b421e0d404b001b7ec